third_party/WebKit/LayoutTests/fast/dom/Range/range-insertNode-separate-endContainer-expected.txt

Test for proper behavior of Range.insertNode(documentFragment) when startContainer != endContainer

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS r.toString() is "ABC"
PASS p.childNodes.length is 4
PASS p.childNodes[0] is t1
PASS p.childNodes[1] is t3
PASS p.childNodes[2] is t4
PASS p.childNodes[3] is t2
PASS r.collapsed is false
PASS r.commonAncestorContainer is p
PASS r.startContainer is p
PASS r.startOffset is 1
PASS r.endContainer is t2
PASS r.endOffset is 3
PASS r.toString() is "PQRXYZABC"
PASS successfullyParsed is true

TEST COMPLETE
